beagle-ui.js: another quick path hack
Using global variable serverBasePath to define where the content for the
site lives. The final solution needs to be pushed across the pages and
made clean. For now, to make many pages work where this variable is not
already defined, I set the default to '../../' since most pages live at
that depth.