HACK _config.yml: don't build static pages on BeagleBone
The static pages are handled by the latest BoneScript server by simply
redirecting accesses to /bone101/static back to /static. This patch
needs to not go to the gh-pages branch, because the static files there
are still needing to be copied from /static to /bone101/static. Moving
forward, we'll simply statically generate the bone101 content and this
patch can be removed.